### 2024-06-06 #6854 @editorialbot
[PRE REVIEW]: AixWeather: A Weather Data Generation Tool for Building Energy System Simulations.  Pull, Transform, Export.
**Submitting author:** @martinraetz (<a href="http://orcid.org/0000-0002-3573-2872">Martin Rätz</a>)
**Repository:** https://github.com/RWTH-EBC/AixWeather
**Branch with paper.md** (empty if default branch): 4-joss_paper
**Version:** v0.1.7
