### 2024-06-06 #6854 @kyleniemeyer
Hi @MartinRaetz, unfortunately we do not have any editors in this area available at the moment, so I have to place this on our waitlist. We will move forward with the process once someone frees up.

In addition, I noticed that your software's license is a modified version of the revised BSD 3-clause license. JOSS submissions must use OSI-approved licenses, and not make modifications to these—please change this to the boilerplate revised BSD3 license. Terms of contributions should go in a separate document (such as CONTRIBUTING.md).
