+++ Joss_Paper/paper.bib
@@ -38,18 +38,6 @@ ABSTRACT = {We present a method to generate Typical Meteorological Year (TMY) da
 DOI = {10.3390/atmos9020053}
 }
 
-
-@misc{EnergyPlus.2017,
-title = {EnergyPlus™, Version 00},
-author = {},
-abstractNote = {Originally developed in 1999, an updated version 8.8.0 with bug fixes was released on September 30th, 2017. EnergyPlus™ is a whole building energy simulation program that engineers, architects, and researchers use to model both energy consumption—for heating, cooling, ventilation, lighting and plug and process loads—and water use in buildings. EnergyPlus is a console-based program that reads input and writes output to text files. It ships with a number of utilities including IDF-Editor for creating input files using a simple spreadsheet-like interface, EP-Launch for managing input and output files and performing batch simulations, and EP-Compare for graphically comparing the results of two or more simulations. Several comprehensive graphical interfaces for EnergyPlus are also available. DOE does most of its work with EnergyPlus using the OpenStudio® software development kit and suite of applications. DOE releases major updates to EnergyPlus twice annually.},
-url = {https://www.osti.gov//servlets/purl/1395882},
-doi = {},
-url = {https://www.osti.gov/biblio/1395882}, year = {2017},
-month = {9},
-note =
-}
-
 @article{EnergyPlus.2001,
 title = {EnergyPlus: creating a new-generation building energy simulation program},
 journal = {Energy and Buildings},
