### 2024-06-21 #6854 @MartinRaetz
Thank you, @kyleniemeyer. I have revised the BSD 3-clause license to the boilerplate version. I have also fixed the failing Bibtex entry.
