### 2024-06-21 #6854 @MartinRaetz
Possible reviewers:
[Jeremy Lerond](https://reviewers.joss.theoj.org/reviewers/3420)
[Florian Kotthoff](https://reviewers.joss.theoj.org/reviewers/2733)
