### 2024-07-25 #6854 @MartinRaetz
Hi @kyleniemeyer, is there anything I can do to help finding an editor?
In my opinion, anyone in simulations using weather data, persons working with weather data, or possibly even persons working with time series data in general might be suitable.
