### 2024-09-10 #6854 @MartinRaetz
Hey @arfon, thanks for taking over!

Here are some suggestions for reviewers:
[Jeremy Lerond](https://reviewers.joss.theoj.org/reviewers/3420)
[Ralph Evins](https://reviewers.joss.theoj.org/reviewers/1254)
[Xuechen Lei](https://reviewers.joss.theoj.org/reviewers/4597)

If you would like me to suggest more, just let me know!
