### 2024-09-21 #6854 @arfon
@ralphevins @leijerry888 – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _AixWeather: A Weather Data Generation Tool for Building Energy System Simulations. Pull, Transform, Export_

Many thanks
Arfon
