### 2024-10-02 487bf24
fix NAN values for DWD
