### 2024-10-04 7471205
add DWD defaults for missing values
