### 2024-10-04 a30511c
added functionality specifying nan ranges
