### 2024-10-05 #6854 @arfon
@FlorianK13 @lymereJ – 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is AixWeather: A Weather Data Generation Tool for Building Energy System Simulations. Pull, Transform, Export

Many thanks
Arfon
