### 2024-10-10 #7344 @editorialbot
[REVIEW]: AixWeather: A Weather Data Generation Tool for Building Energy System Simulations.  Pull, Transform, Export.
**Submitting author:** @martinraetz (<a href="http://orcid.org/0000-0002-3573-2872">Martin Rätz</a>)
**Repository:** https://github.com/RWTH-EBC/AixWeather
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.7
