### 2024-10-22 RWTH-EBC/AixWeather#67 @FlorianK13
[JOSS Review] OSI conform license
This comment is part of https://github.com/openjournals/joss-reviews/issues/7344

Your [license](https://github.com/RWTH-EBC/AixWeather/blob/main/LICENSE) has an added paragraph and hence is not the BSD license anymore. Hence I cannot really check the field of OSI approved licenses in my review checklist.

Is there a reason to have this modification to your license?
