### 2024-11-04 RWTH-EBC/AixWeather#67 @MartinRaetz
@FlorianK13, thank you for the Issue!

I have already updated the license in the [Joss paper branch](https://github.com/RWTH-EBC/AixWeather/tree/4-joss_paper).
My original plan was to merge it to the main branch in one "Joss-PR" once everything is approved. Though, if required I can also make the changes now?
