### 2024-11-05 RWTH-EBC/AixWeather#67 @FlorianK13
Ah great, I didn't see that. Of course you can merge it afterwards as well, I don't mind.
