### 2024-11-07 RWTH-EBC/AixWeather#68 @FlorianK13
[JOSS Review] Invalid DOI formatting
This comment is part of https://github.com/openjournals/joss-reviews/issues/7344

Just a small issue: The editorial bot found this doi with wrong formatting:

❌ INVALID DOIs
