### 2024-11-07 RWTH-EBC/AixWeather#69 @FlorianK13
[JOSS Review] Web-based applications
This comment is part of https://github.com/openjournals/joss-reviews/issues/7344

First of all, I think it is great that you offer AixWeather as a web app at https://aixweather.eonerc.rwth-aachen.de/
Most users are probably not interested in running the code on their own and just want to transform or download data, which is essentially what you provide. Going the extra step by hosting your services publicly is great.

On web-based apps, they write in the [joss review guidelines](https://joss.readthedocs.io/en/latest/submitting.html#a-note-on-web-based-software):
> A note on web-based software
>
> Many web-based research tools are out of scope for JOSS due to a lack of modularity and challenges testing and maintaining the code. Web-based tools may be considered ‘in scope’ for JOSS, provided that they meet one or both of the following criteria: 1) they are are built around and expose a ‘core library’ through a web-based experience (e.g., R/[Shiny](https://www.rstudio.com/products/shiny/) applications) or 2) the web application demonstrates a high-level of rigor with respect to domain modeling and testing (e.g., adopts and implements a design pattern such as [MVC](https://en.wikipedia.org/wiki/Model%E2%80%93view%E2%80%93controller) using a framework such as [Django](https://www.djangoproject.com/)).
> Similar to other categories of submission to JOSS, it’s essential that any web-based tool can be tested easily by reviewers locally (i.e., on their local machine).

I suppose (and correct me if I'm wrong) that you are aiming at category 1 "core library". By splitting the web app code into [this](https://github.com/RWTH-EBC/AixWeather-WebApp) seperate repository one essential step is already done. However, to see AixWeather as a real "core library", I miss these two things:
* Clear installation instructions and a prominent reference to your pypi package
* Some sort of "Getting started" on your documentation page, that guides me through the core process of the package, with examplary code to import, transform, and export weather data
