### 2024-11-12 #7344 @FlorianK13
Hi @arfon, I have already started to create issues directly within the `AixWeather` repository and linked them to this review issue here.
