### 2024-11-15 18ca6c0
delete HTTPS DOI prefix
