+++ Joss_Paper/paper.bib
@@ -47,7 +47,7 @@ pages = {319-331},
 year = {2001},
 note = {Special Issue: BUILDING SIMULATION'99},
 issn = {0378-7788},
-doi = {https://doi.org/10.1016/S0378-7788(00)00114-6},
+doi = {10.1016/S0378-7788(00)00114-6},
 url = {https://www.sciencedirect.com/science/article/pii/S0378778800001146},
 author = {Drury B. Crawley and Linda K. Lawrie and Frederick C. Winkelmann and W.F. Buhl and Y.Joe Huang and Curtis O. Pedersen and Richard K. Strand and Richard J. Liesen and Daniel E. Fisher and Michael J. Witte and Jason Glazer},
 keywords = {Building performance, Simulation, Energy performance, Heat balance, Mass balance, Modular simulation},
