### 2024-11-15 RWTH-EBC/AixWeather#68 @MartinRaetz
Thank you, @FlorianK13.

I have deleted the prefix in commit 18ca6c0ffe78047d6084cbb466592168fee27357.

I do think this will solve the issue. If not, please feel free to reopen.
