### 2024-11-15 f305373
Add Installation guideline and clearer instructions
