+++ README.md
@@ -20,12 +20,13 @@ With the name AixWeather we follow this local tradition.
 
 If you have any questions regarding AixWeather, feel free to contact us at aixweather@eonerc.rwth-aachen.de.
 
+Available at PyPI as [AixWeather](https://pypi.org/project/AixWeather/).
+
 ## WebApp
 
 For fast access without setup, use the WebApp.
 The repository [AixWeather-WebApp](https://github.com/RWTH-EBC/AixWeather-WebApp) contains the corresponding WebApp for this repo.
-This repository contains the functionalities useful for, e.g., automation of tasks and is 
-available at PyPI as [AixWeather](https://pypi.org/project/AixWeather/).
+This repository contains the functionalities useful for, e.g., automation of tasks.
 
 ## How to use
 Clone the repo and simply install via:
