### 2024-11-15 RWTH-EBC/AixWeather#69 @MartinRaetz
Thank you @FlorianK13.

I do think commit f305373022eaf83e7c0ac8f69f427f84a57c9b50 and d2d7337fe39d64cb2043881f93ac673fae71f463 will solve this issue.

Please close the issue if things are clear now.
