+++ examples/e1_pull_DWD_historical_to_all_output_formats.py
@@ -12,14 +12,15 @@ def e1_pull_DWD_historical_to_all_output_formats():
     2. See examplary use of additional features and settings
     3. Create weather data files
     """
-    # Enable logging to see what is happening
+    # Enable logging, this is just get more feedback through the terminal
     import logging
     logging.basicConfig(level="DEBUG")
 
-    # choose the project class according to the desired weather data origin
+    # Choose the project class according to the desired weather data origin.
+    # Check the project classes file or the API documentation to see which classes are available.
     from aixweather.project_class import ProjectClassDWDHistorical
 
-    # initiate the project class which contains or creates all variables and functions
+    # Step 0: Initiate the project class which contains or creates all variables and functions.
     DWD_pull_project = ProjectClassDWDHistorical(
         start=dt.datetime(2022, 1, 1),
         end=dt.datetime(2023, 1, 1),
@@ -30,13 +31,13 @@ def e1_pull_DWD_historical_to_all_output_formats():
         abs_result_folder_path=None,
     )
 
-    # import historical weather from the DWD open access database
+    # Step 1: Import historical weather from the DWD open access database
     DWD_pull_project.import_data()
     print(
         f"\nHow the imported data looks like:\n{DWD_pull_project.imported_data.head()}\n"
     )
 
-    # convert this imported data to the core format
+    # Step 2: Convert this imported data to the core format
     DWD_pull_project.data_2_core_data()
     print(f"\nHow the core data looks like:\n{DWD_pull_project.core_data.head()}\n")
 
@@ -47,7 +48,7 @@ def e1_pull_DWD_historical_to_all_output_formats():
     plot = plot_heatmap_missing_values(DWD_pull_project.core_data)
     plot.show()
 
-    # convert this core data to an output data format of your choice
+    # Step 3: Convert this core data to an output data format of your choice
     DWD_pull_project.core_2_csv()
     DWD_pull_project.core_2_json()
     DWD_pull_project.core_2_pickle()
