### 2024-11-25 RWTH-EBC/AixWeather#69 @FlorianK13
We do agree on the “99% of the users will go for the trivial web app” – as mentioned above I like this web app a lot for its simplicity. It gives most users exactly what they want, ie the transformed weathr data. I also understand your points about maintenance and single source of truth.
For joss however the review focuses on the core software package, its functionalities, modularity, and reusability. A docs page is the core feature to assure reusability. Things that should be answered from the docs page:
-	What data do I get?
-	What are the different transformation processes the package offers?
-	AixWeather ensures data accuracy? What does that mean? In which functions does this happen? What does happen there? Especially for this sentence in your paper I cannot find references in your documentation: “including consideration of time zones, the time reference of the measurement, unit conversions, correct handling of incomplete data series and, where possible, the avoidance of interpolation-related smoothing.”
-	What are typical use cases and workflows where the usage of AixWeather makes my live easier?

Ultimately if I want to reuse AixWeather as a software package in something that I build on my own, all these questions are important and I think they need to be answered. And an example python file is, in my opinion, not enough to achieve that.
