### 2024-11-25 RWTH-EBC/AixWeather#69 @FlorianK13
However, since I’m also rather new as a reviewer for joss, these requirements come from my own interpretation of the joss review docs. Maybe I’m wrong, so I would also be happy to include the other reviewer and/or the editor into this discussion.
