### 2024-11-26 RWTH-EBC/AixWeather#69 @MartinRaetz
Hey @FlorianK13,

thank you for specifying!
I will think about how to integrate such information without generating something that needs regular maintenance.

Here is a commit that would hopefully clarify some of your points: 1234a63c2fc819cf8eb9753a402d65337377222c

Now some explanations regarding your questions:

You asked what data the user gets. Well, I would describe it with "Weather data" for building energy simulations. (I have clarified this in the commit). The input and output formats, as well as the transformation processes, are shown in the figure. What data points these are exactly, will always depend on the desired format and the availability of the data. (I have also clarified this.) The same applies to the transformation processes, if the unit of a data point is "wrong" a transformation to the desired unit is done. Otherwise, no unit transformation is done. So this will always depend on the input data and the chosen output format.

Speaking about the use cases. I do think it's building energy simulations. I mean you might use the weather data for whatever you want to do, though the tool is designed to deliver weather data for building energy simulations.

Regarding the "accuracy", I have adapted the formulation in the paper to make it clearer. I meant, that accuracy is ensured through the consideration of the stated points. Also, these points are not super complex functions. Consideration of time zones just means, that the tool considers (and knows) the time zone of the raw input data and shifts the data accordingly to the desired time zone in the output format. Same simplicity for the units, respecting incomplete data without applying wrong interpolation schemes. So these points are important, and actually often done wrong, but intrinsically trivial.
The interpolation-related smoothing is explained in the figure.

If you want to reuse AixWeathers internal functions, you will need to dive into the code. These functions are under constant extension and improvement. Not only because the tool is supposed to grow and support more data formats, but also cause input formats or output formats might change and the tool needs to be updated to these changes. If you just want to use AixWeathers functionality to get weather data, I think the example file is exactly what you need. You can copy it to your code and you're done.
