### 2024-11-26 RWTH-EBC/AixWeather#69 @FlorianK13
I think our discussion is going in circles. As I see it, the primary goal of AixWeather is the provision of the web service. Creating and maintaining a large documentation page is lots of work, which contrasts the primary goal of providing the web service, as it shifts focus away from work that benefits the web service.

JOSS however is a journal about open source software packages, from and for the research community. If other researchers want to use the core software of AixWeather, they need to know in detail what it provides. Hence, JOSS specifically requests a review of the available documentation. In my opinion, the current docs of AixWeather fail this part of the review.

What I expect from a docs page:
- I want to understand, as quickly as possible, what the provided software can do for me. Is it really what I need? Here it would be nice to have some kind of "statement of need" as the first visible thing at https://rwth-ebc.github.io/AixWeather//main/docs/index.html with the information that I can get DWD weather data and export it in format a, b, c.
- I now see that this is the software I need. How do I integrate it in my code? I maybe want to import from source A, do some of the transformations, and then export to format B. How do I do that? What transformations can I apply, which sources do exist, which exports? Some sort of manual with code and references to the API are needed here.

In the end, you are the maintainer of this software, so you now best how your potential users can use it and what questions might come up for them.

I also checked the last five publications on JOSS and their docs pages: [snk](https://snk.wytamma.com/managing_workflows/), [BrightEyes-MCS](https://github.com/VicidominiLab/BrightEyes-MCS/wiki#input--output-connections), [Phylox](https://phylox.readthedocs.io/en/latest/index.html), [PhenoFeatureFinder](https://phenofeaturefinder.readthedocs.io/en/latest/index.html), [lpjmlkit-1](https://pik-piam.r-universe.dev/lpjmlkit), [lpjmlkit-2](https://pik-piam.r-universe.dev/lpjmlkit/doc/manual.html)
