### 2024-11-26 #7344 @lymereJ
@MartinRaetz - Thank you for your patience and for your submission! I think that this overall is a good product. I do have a few review comments and questions. I am not yet finished with the review but I thought that perhaps I should already share some of my comments/questions to not delay to process further.
- Has the software been used in other publications?
- It's not clear to me what are the roles of the co-authors. Unless I am mistaken, they have not contributed to the repository. Have they been providing guidance on the development?
- It might be good to include a discussion in the paper about how this software compares with similar tools such as for example [Elements](https://bigladdersoftware.com/projects/elements/).
- It seems like there is an issue with the logos used in the documentation, both for this software and Aachen University's.
- I was able to look at the examples and try them, however, I feel like the documentation does not clearly point to the example files on the repository. Additionally, it would be nice to include in the documentation some simple instructions to go along the example scripts.
- I was able to find the package on pypi and install it with `pip` but, unless I am mistaken, I don't see a note/section in the documentation about installation.
