### 2024-11-26 RWTH-EBC/AixWeather#69 @lymereJ
Sorry for including myself in the discussion. I left some preliminary review comments [here](https://github.com/openjournals/joss-reviews/issues/7344) and I see that some of them are echoed here by @FlorianK13. I agree that the documentation should be expanded as mentioned in the comment above. There is always a possibility that funding for hosting the web application dries up. In that case, it is essential to have a good documentation so people can use the tool offline.
