### 2024-11-27 #7344 @MartinRaetz
Hey @lymereJ,

thank you for taking the time to review and also for sharing your thoughts!

1. Yes the software is used (maybe) on a daily basis (at least) by people in our institute. Though, I cannot tell whether it has been referenced in any publication. That is, first, cause it has no citable source yet, and, second, people tend to overlook to reference where they got such files as weather data from.
2. The two authors would be named for "funding acquisition" and "supervision".
3. - [ ] @MartinRaetz: I will have a look at this and will highlight differences in the discussion.
4. Yes, I have already spent quite some time to fix this bug without success. It will either not render on the readme page or on the docs page. Though, as I think the readme is the main source of information and is always available, I decided it would be more important to have the figures there.
5. Please checkout the [Joss-Paper Branch](https://github.com/RWTH-EBC/AixWeather/tree/4-joss_paper) there I have added a hint to the examples. Also, be aware that the docs page won´t be updated unless I push it to the main branch. I would prefer to merge this branch to the main after the review process is finished if that's okay.
6. You will find a section about installation in the Joss-Paper branch.
