### 2024-11-27 RWTH-EBC/AixWeather#69 @MartinRaetz
Thank you @FlorianK13 and @lymereJ,

- [ ] @MartinRaetz: I will have another look at the documentation, with your suggestions and thoughts I hope to come up with something that covers both needs. However, I might need some time to do a deeper dive.

What are your thoughts on shutting down the GitHub docs page and focusing documentation only on the readme (and example file)? The external docs page always comes with quite some pain to maintain, and everything except for the "aixweather" tab with the modules overview, is (or can) be included in the Readme. Also, bear in mind that the docs page you see is only rendered for the current main branch, so changes I do now are not automatically mirrored online.
I also think, that since the figures are not rendering on the docs page, people might overlook the information in the "structure" figure if they decide to read the "readme" on the docs pages and not on the readme page. It also might deliver a wrong image of the tool. E.g. the tool is not meant to be used like e.g. Scikit-Learn where you have many functions that you can combine to your liking to your "own" script. AixWeather is and must be, a streamlined process, always running the same three functions in the same order. None of those internal transformation functions are designed to be used out of that process, the process is designed to choose the correct transformation on its own. And this is, in my opinion, the strength, since this will ensure that you will get the weather data correctly transformed to your desired output format from an often obscure input data format.

I will also try to include these thoughts in the documentation.
