### 2024-11-27 RWTH-EBC/AixWeather#69 @FlorianK13
I'm fine with a documentation in the README
