### 2024-12-02 af0b74d
add example.md to index
