+++ docs/source/examples.rst
@@ -1,1 +1,2 @@
 .. mdinclude::  ../../examples/README.md
+.. mdinclude::  ../../examples/e1_pull_DWD_historical_to_all_output_formats.md
