### 2024-12-02 28ca3f4
Update 'docs' folder with 4-joss_paper-examples-ci and delete old branches.
