+++ docs/4-joss_paper-examples-ci/docs/_sources/examples/e1_pull_DWD_historical_to_all_output_formats.md.txt
@@ -1,17 +1,31 @@
 
 ```python
 """Contains examples"""
+```
 
+pylint: disable=C0415, C0103
+let pylint ignore to import from aixweather (Import outside toplevel), and ignore that pylint
+suggests e1_pull_DWD_historical_to_all_output_formats would not be snake_case.
+
+```python
 import datetime as dt
 ```
 
-choose the project class according to the desired weather data origin
+Enable logging, this is just get more feedback through the terminal
+
+```python
+import logging
+logging.basicConfig(level="DEBUG")
+```
+
+Choose the project class according to the desired weather data origin.
+Check the project classes file or the API documentation to see which classes are available.
 
 ```python
 from aixweather.project_class import ProjectClassDWDHistorical
 ```
 
-initiate the project class which contains or creates all variables and functions
+Step 0: Initiate the project class which contains or creates all variables and functions.
 
 ```python
 DWD_pull_project = ProjectClassDWDHistorical(
@@ -25,7 +39,7 @@ DWD_pull_project = ProjectClassDWDHistorical(
 )
 ```
 
-import historical weather from the DWD open access database
+Step 1: Import historical weather from the DWD open access database
 
 ```python
 DWD_pull_project.import_data()
@@ -34,7 +48,7 @@ print(
 )
 ```
 
-convert this imported data to the core format
+Step 2: Convert this imported data to the core format
 
 ```python
 DWD_pull_project.data_2_core_data()
@@ -54,7 +68,7 @@ plot = plot_heatmap_missing_values(DWD_pull_project.core_data)
 plot.show()
 ```
 
-convert this core data to an output data format of your choice
+Step 3: Convert this core data to an output data format of your choice
 
 ```python
 DWD_pull_project.core_2_csv()
