### 2024-12-02 fdbe76f
use raw image sites
