### 2024-12-03 6a6558f
Update example for CI and doc
