### 2024-12-03 b8a3b7f
Update 'docs' folder with 4-joss_paper-examples-ci and delete old branches.
