+++ docs/4-joss_paper-examples-ci/docs/_sources/examples/e1_pull_DWD_historical_to_all_output_formats.md.txt
@@ -1,5 +1,5 @@
 
-Historical DWD pull example
+# Historical DWD pull example
 
 Enable logging, this is just get more feedback through the terminal
 
