### 2024-12-03 e6d5f9d
Revert "try referencing branch dynamically"
This reverts commit b08f92094e566522fabb8129061eee1ea84f2f7f.
