### 2024-12-03 5be4e70
chore(examples): Automatic commit of example files in Markdown and Jupyter Notebook format.
