+++ docs/source/examples/e1_pull_DWD_historical_to_all_output_formats.md
@@ -1,5 +1,5 @@
 
-# Historical DWD pull example
+# AixWeather Tutorial
 
 Enable logging, this is just get more feedback through the terminal
 
