### 2024-12-03 4d7c57a
Update 'docs' folder with 4-joss_paper and delete old branches.
