+++ README.md
@@ -15,6 +15,12 @@ JOSS-PAPER REVIEW IS FINISHED.
 # AixWeather
 **A weather data generation tool for building energy system simulations. Pull, Transform, Export.**
 
+AixWeather is a streamlined tool which will help you to generate weather data for building energy system simulations.
+The user can choose a data source and an output format, and the tool will take care of the rest.
+Read below to find out what input sources and output formats are currently supported.
+This tool is not designed as a library but as a standalone tool for generating weather 
+data files.
+
 It is being developed at RWTH Aachen University, E.ON Energy Research Center, 
 Institute for Energy Efficient Buildings and Indoor Climate (EBC) in Aachen, Germany.
 
@@ -32,7 +38,7 @@ For fast access without setup, use the WebApp.
 The repository [AixWeather-WebApp](https://github.com/RWTH-EBC/AixWeather-WebApp) contains the corresponding WebApp for this repo.
 This repository contains the functionalities useful for, e.g., automation of tasks.
 
-## How to use
+## How to use AixWeather locally
 Clone the repo and simply install via:
 ```bash
 pip install AixWeather
@@ -43,9 +49,14 @@ cd AixWeather
 pip install -e .
 ```
 
-Then go to examples and follow the descriptions.
+Then go to the [example](https://rwth-ebc.github.io/AixWeather//4-joss_paper//docs/examples/e1_pull_DWD_historical_to_all_output_formats.html) and follow the descriptions.
+This example contains the streamlined process exemplary shown for pulling DWD historical data and exporting it to all available output formats.
+Use generic functions to execute the streamlined process to create the desired output file.
+You may then either run this example file as a standalone and just transfer the output 
+files to your simulation environment or if the weather data shall be generated by means of your own script, 
+just copy the same executive functions to your script.
 For each data origin, a class is defined. 
-Use generic functions to create the desired output file.
+Choose the class according to the desired weather data origin.
 For background knowledge, check the below tool structure and overview.
 
 ## Tool structure and overview
@@ -76,13 +87,25 @@ It retrieves, formats, enriches, and ultimately exports weather data in various 
 Which transformations are executed, i.e. required, will always depend on the input data and the 
 desired output format.
 The current set of implemented transformation functions can be found in 
-[aixweather/transformation_functions](aixweather/transformation_functions).
-The tool selects the required transformation functions based on the input data and 
-the desired output format.
+[aixweather/transformation_functions](https://rwth-ebc.github.
+io/AixWeather//4-joss_paper//docs/code/aixweather.transformation_functions.html), e.g. the [unit 
+conversions](https://rwth-ebc.github.io/AixWeather//4-joss_paper//docs/code/aixweather.
+transformation_functions.html#module-aixweather.transformation_functions.unit_conversions) or 
+the functions to calculate missing values: [variable transformation](https://rwth-ebc.github.
+io/AixWeather//4-joss_paper//docs/code/aixweather.transformation_functions.html#module-aixweather.transformation_functions.variable_transformations).
+Though, these function do not need be called explicitly, as they are called by the tool 
+automatically based on the input data and the desired output format.
+Only if you want to add a new input or output format, you will need to specify which 
+transformations are required specifically for your new format. 
+The community will be happy to have such developments contributed back to the main repository.
+
 
 **Overview**
 
 The below diagram shows the tool's structure and the data flow from the import to the output format.
+It also shows what kind of input and output formats are currently supported.
+Every input is first transformed to the core format, which is then transformed to the desired output format.
+This allows for a clear separation of concerns and a modular structure that is easy to extend.
 ![Weather Data Tool Diagram](https://raw.githubusercontent.com/RWTH-EBC/AixWeather/4-joss_paper/docs/Overview_WeatherTool.png)
 
 **Known limitations:**
