### 2024-12-13 9cb7d3a
Update 'docs' folder with 4-joss_paper and delete old branches.
