+++ docs/4-joss_paper/docs/_sources/examples/e1_pull_DWD_historical_to_all_output_formats.md.txt
@@ -1,4 +1,3 @@
-
 # AixWeather Tutorial
 
 Enable logging, this is just get more feedback through the terminal
