### 2024-12-13 RWTH-EBC/AixWeather#69 @MartinRaetz
Hey @lymereJ and @FlorianK13,

fortunately, we (mainly @FWuellhorst) could fix and improve the CI that creates the docs pages. It is now a lot more stable.

Additionally, I have improved the documentation. There is now a direct explanation of what the tool is and what to expect from it right at the start of the readme and docs pages, and again in the tutorial/example (@FlorianK13 called it statement of need). Also, in the "how to use" section more information is added and the user is forwarded to a tutorial/example available on the docs page. To enable you to check the updated docs pages I have temporarily linked the pages generated by the joss-branch. I will change this back to the main branch upon merging.

When checking, please make sure to be in the https://github.com/RWTH-EBC/AixWeather/tree/4-joss_paper branch.

Thank you for your patience, unfortunately, this review coincides with the last months of my PhD (which is not related to this open-source contribution).

I will now have a look at the things @lymereJ mentioned in the joss-review thread.

Thanks again for volunteering!
