+++ Joss_Paper/paper.bib
@@ -54,3 +54,20 @@ keywords = {Building performance, Simulation, Energy performance, Heat balance, 
 abstract = {Many of the popular building energy simulation programs around the world are reaching maturity — some use simulation methods (and even code) that originated in the 1960s. For more than two decades, the US government supported development of two hourly building energy simulation programs, BLAST and DOE-2. Designed in the days of mainframe computers, expanding their capabilities further has become difficult, time-consuming, and expensive. At the same time, the 30 years have seen significant advances in analysis and computational methods and power — providing an opportunity for significant improvement in these tools. In 1996, a US federal agency began developing a new building energy simulation tool, EnergyPlus, building on development experience with two existing programs: DOE-2 and BLAST. EnergyPlus includes a number of innovative simulation features — such as variable time steps, user-configurable modular systems that are integrated with a heat and mass balance-based zone simulation — and input and output data structures tailored to facilitate third party module and interface development. Other planned simulation capabilities include multizone airflow, and electric power and solar thermal and photovoltaic simulation. Beta testing of EnergyPlus began in late 1999 and the first release is scheduled for early 2001.}
 }
 
+@misc{Meteonorm.2023,
+    author = {Meteotest AG},
+    title = {Meteonorm: Global Climate Data for Solar Applications and Building Design},
+    year = {2023},
+    publisher = {Meteotest AG},
+    url = {https://meteonorm.com/},
+}
+
+@misc{Elements.2016,
+    title = {Elements: Weather File Tool},
+    year = {2016},
+    publisher = {Big Ladder Software},
+    url = {https://bigladdersoftware.com/projects/elements/},
+    organization = {Big Ladder Software}
+}
+
+
