+++ README.md
@@ -86,13 +86,7 @@ It retrieves, formats, enriches, and ultimately exports weather data in various 
 
 Which transformations are executed, i.e. required, will always depend on the input data and the 
 desired output format.
-The current set of implemented transformation functions can be found in 
-[aixweather/transformation_functions](https://rwth-ebc.github.
-io/AixWeather//4-joss_paper//docs/code/aixweather.transformation_functions.html), e.g. the [unit 
-conversions](https://rwth-ebc.github.io/AixWeather//4-joss_paper//docs/code/aixweather.
-transformation_functions.html#module-aixweather.transformation_functions.unit_conversions) or 
-the functions to calculate missing values: [variable transformation](https://rwth-ebc.github.
-io/AixWeather//4-joss_paper//docs/code/aixweather.transformation_functions.html#module-aixweather.transformation_functions.variable_transformations).
+The current set of implemented transformation functions can be found in [aixweather/transformation_functions](https://rwth-ebc.github.io/AixWeather//4-joss_paper//docs/code/aixweather.transformation_functions.html), e.g. the [unit conversions](https://rwth-ebc.github.io/AixWeather//4-joss_paper//docs/code/aixweather.transformation_functions.html#module-aixweather.transformation_functions.unit_conversions) or the functions to calculate missing values: [variable transformation](https://rwth-ebc.github.io/AixWeather//4-joss_paper//docs/code/aixweather.transformation_functions.html#module-aixweather.transformation_functions.variable_transformations).
 Though, these function do not need be called explicitly, as they are called by the tool 
 automatically based on the input data and the desired output format.
 Only if you want to add a new input or output format, you will need to specify which 
