### 2024-12-13 577238b
update doc strings for more clarity
