### 2024-12-16 RWTH-EBC/AixWeather#69 @FlorianK13
The new docs can be accessed directly here, might be useful for review @MartinRaetz @lymereJ  https://rwth-ebc.github.io/AixWeather//4-joss_paper/docs/index.html#
