### 2024-12-16 RWTH-EBC/AixWeather#69 @FlorianK13
The work on the docs page looks good and my major concerns were adressed. The following points are only smaller things:
- I'm missing a link to the actual webapp in your [Readme](https://github.com/RWTH-EBC/AixWeather/tree/4-joss_paper?tab=readme-ov-file#webapp)
- When you write about the [project classes](https://rwth-ebc.github.io/AixWeather//4-joss_paper/docs/examples/e1_pull_DWD_historical_to_all_output_formats.html) in the tutorial, you could link to the [API documentation](https://rwth-ebc.github.io/AixWeather//4-joss_paper/docs/code/aixweather.html#module-aixweather.project_class) of those classes
- I was struggling to get a second example running. Maybe you can extend the tutorial with one more project class, eg the DWD Forecast?
