### 2024-12-18 RWTH-EBC/AixWeather#69 @lymereJ
Thanks @MartinRaetz and @FlorianK13! This looks great.

On the main page of the documentation you have the following:
```
Clone the repo and simply install via:

pip install AixWeather

or

cd AixWeather
pip install -e .
```

I wonder if you shouldn't have the following:
```
Simply install via:

`pip install AixWeather`

Or, by cloning the repository:

`git clone https://github.com/RWTH-EBC/AixWeather.git AixWeather`
`cd AixWeather`
`pip install -e .`
```
