### 2024-12-23 b083665
add examples for projects
