+++ examples/e1_pull_DWD_historical_to_all_output_formats.py
@@ -13,20 +13,21 @@ def e1_pull_DWD_historical_to_all_output_formats():
     # To run the tool for other weather data sources, just change the project class.
     # The rest of the code is streamlined and will remain the same.
 
-
     # Enable logging, this is just to get more feedback through the terminal
     import logging
+
     logging.basicConfig(level="DEBUG")
 
     # Choose the project class according to the desired weather data origin.
-    # Check the project classes file or the API documentation to see which classes are available.
+    # Check the project_class.py file or the API documentation to see which classes are available.
     # [project classes](https://rwth-ebc.github.io/AixWeather//4-joss_paper//docs/code/aixweather.html#module-aixweather.project_class)
     from aixweather.project_class import ProjectClassDWDHistorical
 
     # Step 0: Initiate the project class which contains or creates all variables and functions.
     # For this, we use the datetime module to specify dates.
     import datetime as dt
-    DWD_pull_project = ProjectClassDWDHistorical(
+
+    project = ProjectClassDWDHistorical(
         start=dt.datetime(2022, 1, 1),
         end=dt.datetime(2023, 1, 1),
         station=15000,
@@ -37,28 +38,62 @@ def e1_pull_DWD_historical_to_all_output_formats():
     )
 
     # Step 1: Import historical weather from the DWD open access database
-    DWD_pull_project.import_data()
-    print(
-        f"\nHow the imported data looks like:\n{DWD_pull_project.imported_data.head()}\n"
-    )
+    project.import_data()
+    print(f"\nHow the imported data looks like:\n{project.imported_data.head()}\n")
 
     # Step 2: Convert this imported data to the core format
-    DWD_pull_project.data_2_core_data()
-    print(f"\nHow the core data looks like:\n{DWD_pull_project.core_data.head()}\n")
+    project.data_2_core_data()
+    print(f"\nHow the core data looks like:\n{project.core_data.head()}\n")
 
     # You may optionally also use data quality check utils, like:
     from aixweather.data_quality_checks import plot_heatmap_missing_values
 
     # Plot data quality
-    plot = plot_heatmap_missing_values(DWD_pull_project.core_data)
+    plot = plot_heatmap_missing_values(project.core_data)
     plot.show()
 
     # Step 3: Convert this core data to an output data format of your choice
-    DWD_pull_project.core_2_csv()
-    DWD_pull_project.core_2_json()
-    DWD_pull_project.core_2_pickle()
-    DWD_pull_project.core_2_mos()
-    DWD_pull_project.core_2_epw()
+    project.core_2_csv()
+    project.core_2_json()
+    project.core_2_pickle()
+    project.core_2_mos()
+    project.core_2_epw()
+
+    # By the way, if you want to run this with some other weather data source, here are some copy-paste snippets to get you started:
+
+    # For DWD Forecast data
+    from aixweather.project_class import ProjectClassDWDForecast
+
+    project = ProjectClassDWDForecast(
+        station="06710",  # Example station: Lausanne
+        abs_result_folder_path=None,  # Optional: specify result path
+    )
+
+    # For EPW and TRY we need to import the .epw or .dat file
+    # These imports are just to create the path to the file stored in the tests folder
+    import os
+    from aixweather import definitions
+
+    # For EPW files
+    from aixweather.project_class import ProjectClassEPW
+
+    project = ProjectClassEPW(
+        path=os.path.join(
+            definitions.ROOT_DIR,
+            "tests/test_files/regular_tests/EPW/test_EPW_Essen_Ladybug/input/DEU_NW_Essen_104100_TRY2035_05_Wint_BBSR.epw",
[…]
