+++ docs/source/examples.rst
@@ -3,4 +3,4 @@
 .. toctree::
    :maxdepth: 2
 
-   examples/e1_pull_DWD_historical_to_all_output_formats
+   examples/e1_pull_DWD_historical_to_all_output_formats.md
