### 2024-12-23 de5be26
Update 'docs' folder with 4-joss_paper and delete old branches.
