### 2024-12-23 RWTH-EBC/AixWeather#69 @MartinRaetz
Thank you @FlorianK13 and @lymereJ!

I have addressed your points in the last two commits.

@FlorianK13: I have added additional "copy&paste"-snippets for the other project classes at the end of the tutorial. They work on my computer, let me know if you are having issues!

Thanks again for your time and suggestions!

Have a wonderful Christmas break!
