### 2024-12-23 #7344 @MartinRaetz
Dear @arfon,

is it possible to add another author to the publication?
@FWuellhorst https://github.com/FWuellhorst

He helped a lot with the quality management processes through the CI.
He also made minor content-related PRs and is planning on a larger improvement at the beginning of next year.

Wishing you a Merry Christmas and a Happy New Year!
