### 2024-12-27 #7344 @arfon
> is it possible to add another author to the publication?
> @FWuellhorst https://github.com/FWuellhorst

Yes, this is fine.

Also, @lymereJ, @FlorianK13 – it would be good to hear from you sometime soon about the changes that @MartinRaetz has made thus far and whether they are satisfactory.
