### 2025-01-06 RWTH-EBC/AixWeather#69 @FlorianK13
Looks good so far. I haven't managed to get the EPW and TRY example running. I think the reason is that the
`definitions.ROOT_DIR` links to `.venv/lib/python3.12/site-packages` and not to the actual git repo if you have installed the package via pip. For me it would also be fine to have something like this in the docs:

"To parse EPW or TRY files, you need them locally. You can download example files from our repository at github here: INSERT LINK

Then parse them as follows:

project = ProjectClassEPW(
    path="path/to/your/epw/file"
    )
"
