### 2025-01-06 3222ffd
update info about EPW and TRY
