### 2025-01-06 b16f03a
Clarify the need to have the repo downloaded.
