### 2025-01-06 5694707
clarify need of setting a path
