### 2025-01-06 RWTH-EBC/AixWeather#69 @MartinRaetz
Hey @FlorianK13,

just to verify, did you clone or download the repository and execute the example file from there?
To me the implementation of ROOT_DIR seems to be fine, it should take the location of the executed file as the origin and walk to the ROOT_DIR from there. This should be independent of the way the environment is installed.

I have clarified the requirements as you suggested.
