### 2025-01-06 RWTH-EBC/AixWeather#69 @FlorianK13
Yes, it should work if people download the repo and install the software from there. But if people get software by installing it via pip, then the ROOT-DIR links to the `site-package` of the python environment - at least that is my assumption of this error.

Have you tried to install the software in a clean environment via pip and then run the example script from there?
