### 2025-01-06 RWTH-EBC/AixWeather#69 @MartinRaetz
Yes, downloading the repo to run the example is required (if you want to use files from the repo). I have clarified this in the comments. You may check the current version of the example. It will upload to the Docs Pages in the next 20 minutes or so.
