### 2025-01-06 RWTH-EBC/AixWeather#69 @MartinRaetz
Thank you @FlorianK13.

 To speed up the process, here is what will be uploaded once the CI is finished:

For EPW and TRY we need to import the .epw or .dat file, e.g. from your local drive. The
Readme provides information on where you can download your own EPW and TRY files. For this
example, we use the test files provided in the tests folder (be sure to have the full
repository downloaded to have access to these files).
If you use your own files, you can specify the path to the file in the `path` argument.

Hope this clarifies the usage?
Thanks for suggesting!
