### 2025-01-06 RWTH-EBC/AixWeather#69 @MartinRaetz
Info: The CI is finished and the doc-pages are updated.
