### 2025-01-06 #7344 @FlorianK13
Hi @arfon. From my site, the package as it is on the `4-joss_paper` branch is fully reviewed now. Thanks to @MartinRaetz for working on all the Issues I opened in the repo. I think that `AixWeather` can be published in JOSS.
