### 2025-01-06 20c441c
add Wüllhorst as author for his help and contributions
