### 2025-01-06 #7344 @MartinRaetz
Thank you for your review and the valuable suggestions to improve the documentation, @FlorianK13!

@arfon: As announced, I have added @FWuellhorst https://github.com/FWuellhorst to the list of authors for the described contributions.
