### 2025-02-24 c177122
delete reminder text
