### 2025-02-24 c4e4c5b
Update 'docs' folder with 4-joss_paper and delete old branches.
