### 2025-02-24 f1b6ec9
4 joss paper (#71)
This PR introduces the JOSS (Journal of Open Source Software) paper submission branch, focusing on enhanced documentation and the manuscript for publication.


* start joss paper

* paper

* paper updates

* github action to create paper

* fix refs

* fix path

* test

* blank

* blank

* some updates on the paper

* paper updates

* .

* .

* .

* .

* .

* .

* .

* grammarly

* .

* .

* drawio update

* .

* .

* update ref energyplus

* Update README.md

* Improve resolution of image

* update LICENSE

* delete unused faulty duplicate reference

* delete HTTPS DOI pref
[…]
