+++ Joss_Paper/paper.bib
@@ -1,0 +1,73 @@
+@article{Maier.2023,
+author = {Laura Maier and David Jansen and Fabian Wüllhorst and Martin Kremer and Alexander Kümpel and Tobias Blacha and Dirk Müller},
+title = {AixLib: an open-source Modelica library for compound building energy systems from component to district level with automated quality management},
+journal = {Journal of Building Performance Simulation},
+volume = {0},
+number = {0},
+pages = {1-24},
+year = {2023},
+publisher = {Taylor & Francis},
+doi = {10.1080/19401493.2023.2250521},
+URL = {https://doi.org/10.1080/19401493.2023.2250521},
+eprint = {https://doi.org/10.1080/19401493.2023.2250521}
+}
+
+@Article{WetterZuoNouiduiPang.2014,
+  author  =  {Michael Wetter and Wangda Zuo and Thierry S. Nouidui and Xiufeng Pang},
+  title   =  {Modelica {Buildings} library},
+  journal =  {Journal of Building Performance Simulation},
+  volume  =  {7},
+  number  =  {4},
+  pages   =  {253--270},
+  year    =  {2014},
+  doi     =  {10.1080/19401493.2013.765506},
+  url     = "https://doi.org/10.1080/19401493.2013.765506"
+}
+
+@Article{PVGIS.2023,
+AUTHOR = {Huld, Thomas and Paietta, Elena and Zangheri, Paolo and Pinedo Pascua, Irene},
+TITLE = {Assembling Typical Meteorological Year Data Sets for Building Energy Performance Using Reanalysis and Satellite-Based Data},
+JOURNAL = {Atmosphere},
+VOLUME = {9},
+YEAR = {2018},
+NUMBER = {2},
+ARTICLE-NUMBER = {53},
+URL = {https://www.mdpi.com/2073-4433/9/2/53},
+ISSN = {2073-4433},
+ABSTRACT = {We present a method to generate Typical Meteorological Year (TMY) data sets for use in calculations of the energy performance of buildings, based on satellite derived solar radiation data and other meteorological parameters obtained from reanalysis products. The great advantage of this method is the availability of data over large geographical regions, giving global coverage for the reanalysis and continental-scale coverage for the solar radiation data, making it possible to generate TMY data for nearly any location, independent of the availability of meteorological measurement stations in the area. The TMY data generated with this method have been validated against 487 meteorological stations in Europe, by calculating heating and cooling degree days, and by running building energy performance simulations using EnergyPlus. Results show that the generated data sets using a long time series perform better than the TMY data generated from station measurements for building heating calculations and nearly as well for cooling calculations, with relative standard deviations remaining below 6% for heating calculations. TMY data constructed using the proposed method yield somewhat larger deviations compared to TMY data constructed from station data. We outline a number of possibilities for further improvement using data sets that will become available in the near future.},
+DOI = {10.3390/atmos9020053}
+}
+
+@article{EnergyPlus.2001,
+title = {EnergyPlus: creating a new-generation building energy simulation program},
+journal = {Energy and Buildings},
+volume = {33},
+number = {4},
+pages = {319-331},
+year = {2001},
+note = {Special Issue: BUILDING SIMULATION'99},
+issn = {0378-7788},
+doi = {10.1016/S0378-7788(00)00114-6},
+url = {https://www.sciencedirect.com/science/article/pii/S0378778800001146},
+author = {Drury B. Crawley and Linda K. Lawrie and Frederick C. Winkelmann and W.F. Buhl and Y.Joe Huang and Curtis O. Pedersen and Richard K. Strand and Richard J. Liesen and Daniel E. Fisher and Michael J. Witte and Jason Glazer},
+keywords = {Building performance, Simulation, Energy performance, Heat balance, Mass balance, Modular simulation},
[…]
