+++ .gitlab-ci.yml
@@ -2,7 +2,7 @@ stages:
  - test
  - doc
  - code_quality
-# - examples
+ - examples
  - build
  - release
  - deploy
@@ -29,11 +29,11 @@ include:
     file: 'pages/gh-pages.gitlab-ci.yml'
   - project: 'EBC/EBC_all/gitlab_ci/templates'
     file: 'python/tests/tests.gitlab-ci.yml'
-#  - project: 'EBC/EBC_all/gitlab_ci/templates'
-#    file: 'python/tests/coverage.gitlab-ci.yml'
+  - project: 'EBC/EBC_all/gitlab_ci/templates'
+    file: 'python/tests/coverage.gitlab-ci.yml'
   - project: 'EBC/EBC_all/gitlab_ci/templates'
     file: 'python/pypi-release/release.gitlab-ci.yml'
-#  - project: 'EBC/EBC_all/gitlab_ci/templates'
-#    file: 'python/examples/examples.gitlab-ci.yml'
+  - project: 'EBC/EBC_all/gitlab_ci/templates'
+    file: 'python/examples/examples.gitlab-ci.yml'
   - template: Dependency-Scanning.gitlab-ci.yml
   - template: SAST.gitlab-ci.yml
