+++ aixweather/__init__.py
@@ -11,4 +11,4 @@ if Version(pd.__version__) >= Version("2.2"):
         " consider downgrading to 2.1"
     )
 
-__version__ = "0.1.7"
+__version__ = "0.1.8"
