+++ docs/main/docs/_sources/examples.rst.txt
@@ -1,1 +1,6 @@
 .. mdinclude::  ../../examples/README.md
+
+.. toctree::
+   :maxdepth: 2
+
+   examples/e1_pull_DWD_historical_to_all_output_formats.md
