### 2025-03-01 #7344 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
